requirements.txt
git+https://github.com/davfsa/hikari@5c1b5e7
hikari-lightbulb==2.2.4
hikari-miru==1.1.2
asyncpg~=0.26.0
aiohttp~=3.8.1
requests==2.28.1
pyYAML~=6.0
black~=22.6.0
